fixed logical bug - should only buy shares where there is enough money to coverbrokerage and the cost of a single share (no partial shares) updated all tests to reflect the fundamental change

+++ MarketAnalysis/Strategy/Parameters/ClusteringParameters.cs
@@ -6,7 +6,7 @@ namespace MarketAnalysis.Strategy.Parameters
 {
     public class ClusteringParameters : IParameters
     {
-        public decimal Threshold { get; set; }
+        public int Threshold { get; set; }
         public int Partitions { get; set; } = 5;
         public Grid<List<decimal>> Grid { get; set; } = new();
         public TimeSpan? OptimisePeriod { get; } = TimeSpan.FromDays(1024);
